Verification of reconfigure unattended-upgrades step
Added a sub-step for the verification of reconfigure unattended-upgrades step.

Adding this step due to opportunity for confusion in the video at the 2:41 mark where TT indicated to select 'no' followed by indicating to select 'y' on the same step. It was probably intended that the 'no' direction would have been edited out, however, it remained in the video...  I felt it important to verify the configuration and had to look up the file location.