example portainer URL
I presumed portainer would be on https vs http.  Because of unrelated difficulties I figured I did something wrong and tried and tried to fix what was not broken. In the video you can't see the URL in TT's browser so there is no indication that you are required to use http in conjuction with the correct port, 30777.