Update ipWhiteList to correct RFC1918 networks (#49)
The 172.16.x.x network is only from 172.16.x.x - 172.31.x.x. I've updated this in the whitelist to ensure this only uses the RFC 1918 networks
The 10.x.x.x. network is a /8 as per the RFC-1918 standards. I've updated this as well
(i'm going on the assumptoin that we want this as private networks and not publicly routable networks)